requirements.txt
torch
sentence-transformers
faiss-gpu  
transformers
accelerate